pyproject.toml
[tool.poetry]
name = "yt2mob"
version = "0.1.0"
description = ""
authors = ["Satanowski <satanowski@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.28.1"
peewee = "^3.15.4"
click = "^8.1.3"
feedparser = "^6.0.10"
loguru = "^0.6.0"
youtube_dl = "^2021.12.17"
cloudpathlib = { version = "^0.10.0", extras = ["s3"] }
PyYAML = "^6.0"
PyDrive2 = "^1.15.0"

[tool.poetry.dev-dependencies]
bpython = "^0.23"
pylint = "^2.15.7"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
